+++ facades/Logger.d.ts
@@ -1,9 +1,9 @@
 import LoggerBuilder from "../builders/LoggerBuilder";
 export default class Logger {
     static setContext(context: string): LoggerBuilder;
-    static debug(value: string): void;
-    static error(value: string): void;
-    static info(value: string): void;
-    static warn(value: string): void;
+    static debug(value: string): LoggerBuilder;
+    static error(value: string): LoggerBuilder;
+    static info(value: string): LoggerBuilder;
+    static warn(value: string): LoggerBuilder;
     static separator(): void;
 }